+++ actionview/test/template/dependency_tracker_test.rb
@@ -59,6 +59,21 @@ class ERBTrackerTest < Minitest::Test
     assert_equal ["messages/message123"], tracker.dependencies
   end
 
+  def test_dependency_of_template_partial_with_layout
+    skip # FIXME: Needs to be fixed properly, right now we can only match one dependency per line. Need multiple!
+    template = FakeTemplate.new("<%# render partial: 'messages/show', layout: 'messages/layout' %>", :erb)
+    tracker = make_tracker("multiple/_dependencies", template)
+
+    assert_equal ["messages/layout", "messages/show"], tracker.dependencies
+  end
+
+  def test_dependency_of_template_layout_standalone
+    template = FakeTemplate.new("<%# render layout: 'messages/layout' do %>", :erb)
+    tracker = make_tracker("messages/layout", template)
+
+    assert_equal ["messages/layout"], tracker.dependencies
+  end
+
   def test_finds_dependency_in_correct_directory
     template = FakeTemplate.new("<%# render(message.topic) %>", :erb)
     tracker = make_tracker("messages/_message", template)